requirements.txt
amqp==5.3.1
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.1
Automat==25.4.16
billiard==4.2.2
celery==5.5.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
Django==5.2.7
django-celery-beat==2.8.1
django-timezone-field==7.1
django_celery_results==2.6.0
h11==0.16.0
httptools==0.7.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
kombu==5.5.4
msgpack==1.1.2
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg==3.2.11
psycopg-binary==3.2.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.3
redis==6.4.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
zope.interface==8.0.1